feat: Add vector drawable for "more" icon
This commit introduces a new vector drawable resource named `more.xml`.

The drawable represents a "more" or icon, typically depicted as three horizontal dots. It is defined with a size of 24dp x 24dp and uses black as the fill color.

This vector drawable provides a scalable and resolution-independent icon for use in the application's UI.